SPlit out typeboxBrand
