+++ packages/dds/tree/src/util/brand.ts
@@ -3,7 +3,6 @@
  * Licensed under the MIT License.
  */
 
-import { NumericOptions, TUnsafe, Type } from "@sinclair/typebox";
 import { UsageError } from "@fluidframework/telemetry-utils";
 import { Covariant, isAny } from "./typeCheck.js";
 import { Assume } from "./typeUtils.js";
@@ -207,25 +206,4 @@ export function brandErased<T extends BrandedType<any, string>>(
 	value: isAny<ValueFromBranded<T>> extends true ? never : ValueFromBranded<T>,
 ): ErasedType<NameFromBranded<T>> {
 	return value as ErasedType<NameFromBranded<T>>;
-}
-
-/**
- * Create a TypeBox string schema for a branded string type.
- * This only validates that the value is a string,
- * and not that it came from the correct branded type (that information is lost when serialized).
- */
-export function brandedStringType<T extends string>(): TUnsafe<T> {
-	// This could use:
-	// return TypeSystem.CreateType<T>(name, (options, value) => typeof value === "string")();
-	// Since there isn't any useful custom validation to do and
-	// TUnsafe is documented as unsupported in `typebox/compiler`,
-	// opt for the compile time behavior like the above, but the runtime behavior of the built in string type.
-	return Type.String() as unknown as TUnsafe<T>;
-}
-
-export function brandedNumberType<T extends number>(
-	options?: NumericOptions<number> | undefined,
-): TUnsafe<T> {
-	// See comments on `brandedStringType`.
-	return Type.Number(options) as unknown as TUnsafe<T>;
 }
